feat(event-display): separate Phoenix menu UI from UI manager
